1. Emit types 2. Moved some typedefs out of a single file into individual files.
