+++ src/markdown/extensions/MarkedExtensionAsPerDocs.ts
@@ -1,6 +1,4 @@
-export type GetPostData = (args: {post?: string, property?: string}) => string|void;
-
-export type OnShortCode = (args: Record<string, unknown>) => Promise<string>;
+import { SCToken } from "./SCToken";
 
 /**
  * The Docs don't seem to match the MarkedExtension type, or I just don't
@@ -17,13 +15,4 @@ export interface MarkedExtensionAsPerDocs {
   async: boolean;
   tokenizer(src: string, tokens: SCToken[]): void;
   renderer(token: SCToken): string;
-}
-
-export interface SCToken {
-  type: string;
-  raw: string;
-  text: string;
-  tokens?: SCToken[];
-  html: string;
-  args: Record<string, unknown>;
 }
