+++ src/markdown/extensions/OnShortCode.ts
@@ -1,0 +1,3 @@
+import { Patrika } from "../../Patrika";
+
+export type OnShortCode = (args: Record<string, unknown>, patrika: Patrika) => Promise<string>;
