+++ src/markdown/extensions/SCToken.ts
@@ -1,0 +1,8 @@
+export interface SCToken {
+  type: string;
+  raw: string;
+  text: string;
+  tokens?: SCToken[];
+  html: string;
+  args: Record<string, unknown>;
+}
