+++ src/markdown/extensions/index.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
 import { getExtension } from "./inline";
-import { MarkedExtensionAsPerDocs } from "./typedefs";
+import { MarkedExtensionAsPerDocs } from "./MarkedExtensionAsPerDocs";
 
 export function getExtensions (): MarkedExtensionAsPerDocs[] {
   return [getExtension()];
