+++ tsconfig.json
@@ -6,6 +6,7 @@
     "lib": ["ES6"],
     "moduleResolution": "Node16",
     "typeRoots": ["./node_modules/@types", "./types"],
+    "declaration": true
   },
   "include": [
     "./src/**/*"
